requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
CacheControl==0.14.4
cachelib==0.13.0
cachetools==6.2.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.3
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
firebase_admin==7.1.0
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Limiter==4.0.0
Flask-Login==0.6.3
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.2
mysql-connector-python==9.4.0
ordered-set==4.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==7.0.0
requests==2.32.3
rich==14.2.0
rq==2.6.1
rsa==4.9.1
six==1.17.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
